kadai-login1.1

+++ app/Http/Controllers/TasksController.php
@@ -6,6 +6,8 @@ use Illuminate\Http\Request;
 
 use App\Task;
 
+use App\Http\Controllers\Controller;
+
 class TasksController extends Controller
 {
     /**
@@ -15,11 +17,16 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+    
+            return view('tasks.index', [
+                'tasks' => $tasks,
+            ]);
+        }else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -29,11 +36,16 @@ class TasksController extends Controller
      */
     public function create()
     {
-        $task = new Task;
-        
-        return view('tasks.create',[
-            'task' => $task,
-            ]);
+        if (\Auth::check()) {
+            $task = new Task;
+            
+            return view('tasks.create',[
+                'task' => $task,
+                ]);
+                
+        }else {
+            return redirect('login');
+        }
     }
 
     /**
@@ -44,16 +56,22 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'status' => 'required|max:10',
-            'content' => 'required|max:191',
+        if (\Auth::check()) {
+            $this->validate($request, [
+                'status' => 'required|max:10',
+                'content' => 'required|max:191',
+                ]);
+                
+            $request->user()->tasks()->create([
+                'status' => $request->status,
+                'content' => $request->content,
             ]);
             
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-        return redirect('/');
+            return redirect('/');
+        
+        }else {
+            return redirect('login');
+        }
     }
 
     /**
@@ -64,11 +82,20 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Task::find($id);
-        
-        return view('tasks.show', [
-            'task' => $task,
-            ]);
+        if (\Auth::check()) {
+            $task = Task::find($id);
+                
+                if (\Auth::id() == $task->user_id) {
+                    return view('tasks.show', [
+                        'task' => $task,
+                        ]);
+                } else {
+                    return redirect('/');
+                }
+            
+        }else {
+            return redirect('login');
+        }
     }
 
     /**
@@ -79,11 +106,20 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        $task = Task::find($id);
-        
-        return view('tasks.edit', [
-            'task' => $task,
-            ]);
+        if (\Auth::check()) {
+            $task = Task::find($id);
+            
+                if(\Auth::id() == $task->user_id) {
+                    return view('tasks.edit', [
+                        'task' => $task,
+                     ]);
+                } else {
+                    return redirect('/');
+                }
+                
+        } else {
+            return redirect('login');
+        }
         
     }
 
@@ -96,17 +132,26 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
-            'status' => 'required|max:10',
-            'content' => 'required|max:191',
-            ]);
-        
-        $task = Task::find($id);
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-        
-        return redirect('/');
+        if (\Auth::check()) {
+            $task = Task::find($id);
+            if(\Auth::id() == $task->user_id) {
+                $this->validate($request, [
+                    'status' => 'required|max:10',
+                    'content' => 'required|max:191',
+                    ]);
+                
+                $task->status = $request->status;
+                $task->content = $request->content;
+                $task->save();
+                
+                return redirect('/');
+            } else {
+                return redirect('/');
+            }
+            
+        } else {
+            return redirect('login');
+        }
     }
 
     /**
@@ -117,9 +162,17 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
-        
-        return redirect('/');
+        if (\Auth::check()) {
+            $task = Task::find($id);
+            if (\Auth::id() == $task->user_id) {
+                $task->delete();
+                return redirect('/');
+            } else {
+                return redirect('/');
+            }
+                
+        } else {
+            return redirect('login');
+        }
     }
 }
